research_paper code/requirements.txt
opencv-contrib-python-headless==4.5.5.64
-f https://download.pytorch.org/whl/cpu/torch_stable.html 
torch==1.10.2+cpu
Pillow==9.1.0
torchvision==0.11.3
tqdm==4.62.3
